Add files via upload
The study of human learning continues to develop and expand. In institutions of higher education, the issue of utilizing modern information and communication technologies for teaching and learning is very important. It uses the latest technologies to assist and enhance knowledge distribution, and calls for flexible and active interactions amongst online teachers and students. The objective of this paper is to build an educational society and support sharing knowledge, Increasing student efficiency by sharing educational content without restriction, Giving students chance to learn some skills etc.